requirements.txt
streamlit
anthropic
spacy
PyPDF2
python-docx
pandas
fpdf
python-dotenv